pyproject.toml
[project]
name = "clinic-appointments"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikita Kuznetsov",email = "nkuz1509@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "greenlet (>=3.2.3,<4.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "anyio (>=4.9.0,<5.0.0)",
    "pytest-asyncio (>=1.0.0,<2.0.0)",
    "pytest-tornasync (>=0.6.0.post2,<0.7.0)",
    "pytest-trio (>=0.8.0,<0.9.0)",
    "pytest-twisted (>=1.14.3,<2.0.0)",
    "twisted (>=25.5.0,<26.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)"
]

[tool.poetry]
packages = [{include = "clinic_appointments", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
